LayoutTests/fast/page-color-sampling/color-sampling-includes-sides-that-had-insets.html

<!DOCTYPE html> <!-- webkit-test-runner [ ContentInsetBackgroundFillEnabled=true pageTopColorSamplingEnabled=true useFlexibleViewport=true ] -->
<html>
<head>
    <meta charset="utf-8">
    <meta name="viewport" content="width=device-width, initial-scale=1">
    <style>
        body, html {
            width: 100%;
            height: 100%;
            margin: 0;
            padding: 0;
            font-family: system-ui;
        }

        .content {
            width: 100%;
            height: 2000px;
            padding: 120px 20px;
            box-sizing: border-box;
        }

        .top {
            position: fixed;
            top: 0;
            left: 0;
            right: 0;
            height: 80px;
            background-color: rgb(0, 122, 255);
            z-index: 100;
        }

        .bottom {
            position: fixed;
            bottom: 0;
            left: 0;
            right: 0;
            height: 80px;
            background-color: rgb(255, 59, 48);
            z-index: 100;
        }
    </style>
    <script src="../../resources/js-test.js"></script>
    <script src="../../resources/ui-helper.js"></script>
    <script>
    jsTestIsAsync = true;

    addEventListener("load", async () => {
        await UIHelper.setObscuredInsets(40, 0, 40, 0);
        await UIHelper.ensurePresentationUpdate();

        initialEdgeColors = await UIHelper.fixedContainerEdgeColors();
        shouldBeEqualToString("initialEdgeColors.top", "rgb(0, 122, 255)");
        shouldBeEqualToString("initialEdgeColors.bottom", "rgb(255, 59, 48)");

        await UIHelper.setObscuredInsets(40, 0, 0, 0);
        await UIHelper.ensurePresentationUpdate();
        afterRemovingInset = await UIHelper.fixedContainerEdgeColors();
        shouldBeEqualToString("afterRemovingInset.top", "rgb(0, 122, 255)");
        shouldBeEqualToString("initialEdgeColors.bottom", "rgb(255, 59, 48)");

        document.querySelector(".bottom").remove();
        await UIHelper.ensurePresentationUpdate();
        afterRemovingBottomContainer = await UIHelper.fixedContainerEdgeColors();
        shouldBeEqualToString("afterRemovingBottomContainer.top", "rgb(0, 122, 255)");
        shouldBeNull("afterRemovingBottomContainer.bottom");

        finishJSTest();
    });
    </script>
</head>
<body>
    <div class="top"></div>
    <div class="bottom"></div>
    <div class="content"></div>
</body>
</html>
